extract client and project setup to e2e test utils
